Now app completely uses memcache.  Also fixed a minor tag issue
git-svn-id: https://subversion.assembla.com/svn/shopmondays@16 68e3252c-69b3-4b82-877e-b3103f299bcd

+++ templatetags/custom.py
@@ -8,7 +8,7 @@ def getkey(value, arg):
 
 def balance( history ):
 	""" returns credit-debit """
-	return str( int(history["money earned"]) - int(history["money spent"]) )
+	return str( float(history["money earned"]) - float(history["money spent"]) )
 
 def dollars( amount ):
 	""" converts a float to dollars """
